refactoring for doc generation. cool. cool cool cool.
