+++ src/SkillRunner/bot/exceptions.py
