+++ src/SkillRunner/bot/storage.py
@@ -3,7 +3,7 @@ import json
 import requests
 import logging
 
-from .apiclient import ApiClient
+from . import apiclient
 
 class Brain(object):
     """
@@ -18,7 +18,7 @@ class Brain(object):
             self.verify_ssl = False
         else:
             self.verify_ssl = True
-        self.api_client = ApiClient(self.request_uri, user_id, api_token, timestamp)
+        self.api_client = apiclient.ApiClient(self.request_uri, user_id, api_token, timestamp)
 
 
     def make_uri(self, key):
